[FIX JENKINS-36055] Move latestRuns data module to "data/runs"
