+++ blueocean-dashboard/src/test/js/run-details-changes-spec.js
@@ -2,7 +2,7 @@ import { assert } from 'chai';
 import React from 'react';
 import sd from 'skin-deep';
 
-import { latestRuns } from './latestRuns';
+import { latestRuns } from './data/runs/latestRuns';
 import RunDetailsChanges from '../../main/js/components/RunDetailsChanges';
 
 describe('RunDetailsChanges', () => {
